refactor(data): adapt and simplify ListMetadataFormatsHandler
- Migrate from OAICompiledRequest to Request
- Reduce complexity
