+++ xoai-data-provider/src/main/java/io/gdcc/xoai/dataprovider/handlers/ListMetadataFormatsHandler.java
@@ -8,61 +8,67 @@
 
 package io.gdcc.xoai.dataprovider.handlers;
 
-import io.gdcc.xoai.dataprovider.exceptions.HandlerException;
 import io.gdcc.xoai.dataprovider.exceptions.InternalOAIException;
-import io.gdcc.xoai.dataprovider.exceptions.NoMetadataFormatsException;
-import io.gdcc.xoai.dataprovider.exceptions.OAIException;
-import io.gdcc.xoai.dataprovider.handlers.helpers.ItemRepositoryHelper;
+import io.gdcc.xoai.dataprovider.exceptions.handler.HandlerException;
+import io.gdcc.xoai.dataprovider.exceptions.handler.NoMetadataFormatsException;
 import io.gdcc.xoai.dataprovider.model.Context;
 import io.gdcc.xoai.dataprovider.model.ItemIdentifier;
 import io.gdcc.xoai.dataprovider.model.MetadataFormat;
-import io.gdcc.xoai.dataprovider.parameters.OAICompiledRequest;
+import io.gdcc.xoai.dataprovider.repository.ItemRepository;
 import io.gdcc.xoai.dataprovider.repository.Repository;
-import io.gdcc.xoai.model.oaipmh.ListMetadataFormats;
+import io.gdcc.xoai.model.oaipmh.Request;
+import io.gdcc.xoai.model.oaipmh.ResumptionToken;
+import io.gdcc.xoai.model.oaipmh.verbs.ListMetadataFormats;
 
+import java.util.ArrayList;
 import java.util.List;
+import java.util.Optional;
 
 
 public class ListMetadataFormatsHandler extends VerbHandler<ListMetadataFormats> {
-    private final ItemRepositoryHelper itemRepositoryHelper;
-
+    private final ItemRepository itemRepository;
+    
     public ListMetadataFormatsHandler(Context context, Repository repository) {
         super(context, repository);
-        itemRepositoryHelper = new ItemRepositoryHelper(repository.getItemRepository());
+        itemRepository = repository.getItemRepository();
 
         // Static validation
         if (getContext().getMetadataFormats() == null ||
                 getContext().getMetadataFormats().isEmpty())
             throw new InternalOAIException("The context must expose at least one metadata format");
     }
-
+    
+    @Override
+    public ListMetadataFormats handle(Request request, ResumptionToken.Value token) throws HandlerException {
+        throw new InternalOAIException("Method ListMetadataFormatsHandler.handle not allowed with resumption token");
+    }
 
     @Override
-    public ListMetadataFormats handle(OAICompiledRequest params) throws OAIException, HandlerException {
-        ListMetadataFormats result = new ListMetadataFormats();
-
-        if (params.hasIdentifier()) {
-            ItemIdentifier item = itemRepositoryHelper.getItem(params.getIdentifier());
-            List<MetadataFormat> metadataFormats = getContext().formatFor(getRepository().getFilterResolver(), item);
-            if (metadataFormats.isEmpty())
-                throw new NoMetadataFormatsException();
-            for (MetadataFormat metadataFormat : metadataFormats) {
-                io.gdcc.xoai.model.oaipmh.MetadataFormat format = new io.gdcc.xoai.model.oaipmh.MetadataFormat()
-                    .withMetadataPrefix(metadataFormat.getPrefix())
-                    .withMetadataNamespace(metadataFormat.getNamespace())
-                    .withSchema(metadataFormat.getSchemaLocation());
-                result.withMetadataFormat(format);
-            }
+    public ListMetadataFormats handle(Request request) throws HandlerException {
+        final List<MetadataFormat> formats = new ArrayList<>();
+        
+        // This verb allows to specify an identifier to return the available metadata formats for this particular item
+        Optional<String> identifier = request.getIdentifier();
+        if (identifier.isPresent()) {
+            ItemIdentifier item = itemRepository.getItem(identifier.get());
+            
+            // Lookup the formats available from the context. When the metadata formats registered within the context
+            // do not define any condition, this will be a simple copy of all formats. When items might have not every
+            // format available, it is up to the implementing application to provide a suiting condition to lookup
+            // supported formats for the item!
+            formats.addAll(getContext().formatFor(item));
         } else {
-            for (MetadataFormat metadataFormat : getContext().getMetadataFormats()) {
-                io.gdcc.xoai.model.oaipmh.MetadataFormat format = new io.gdcc.xoai.model.oaipmh.MetadataFormat()
-                        .withMetadataPrefix(metadataFormat.getPrefix())
-                        .withMetadataNamespace(metadataFormat.getNamespace())
-                        .withSchema(metadataFormat.getSchemaLocation());
-                result.withMetadataFormat(format);
-            }
+            formats.addAll(getContext().getMetadataFormats());
         }
-
+        
+        // If no formats could be found, send an error message
+        if (formats.isEmpty()) {
+            throw new NoMetadataFormatsException();
+        }
+    
+        // Create the response
+        ListMetadataFormats result = new ListMetadataFormats();
+        formats.forEach(format -> result.withMetadataFormat(format.toOAIPMH()));
         return result;
     }
